Extract MischievousUrlDownloader
To allow usage in other tests.
